Update manifest to support reftest chains.
This adds enough information in the manifest to support reftest chains
of the form A == B == (C | D). Followup commits will enforce more of
the required properties for these chains e.g. restrictions on what
kinds of cycle can exist.

As aprt of this the manifest code was refactored, and some other code
was moved around to match.
