+++ .travis.yml
@@ -2,4 +2,4 @@ language: python
 python:
   - "2.7"
 # command to run tests, e.g. python setup.py test
-script:  python tools/scripts/lint.py
+script:  ./lint
