+++ README.md
@@ -13,7 +13,8 @@ need a copy of OpenSSL. For users on Windows this is available from
 [the openssl website](https://www.openssl.org/related/binaries.html).
 
 To get the tests running, you need to set up the test domains in your
-[`hosts` file](http://en.wikipedia.org/wiki/Hosts_%28file%29%23Location_in_the_file_system). The following entries are required:
+[`hosts` file](http://en.wikipedia.org/wiki/Hosts_%28file%29%23Location_in_the_file_system). The
+following entries are required:
 
 ```
 127.0.0.1	web-platform.test
@@ -34,7 +35,7 @@ git submodule update --init --recursive
 The test environment can then be started using
 
 ```
-python serve.py
+./serve
 ```
 
 This will start HTTP servers on two ports and a websockets server on
@@ -70,41 +71,16 @@ convenient way to run the web-platform tests in-browser. It will run
 testharness.js tests automatically but requires manual work for
 reftests and manual tests.
 
-In order to use the runner, it is first necessary to generate a test
-manifest. This must be called `MANIFEST.json` and placed in the
-web-platform-tests root.
-
-You must do this step to use the test runner, even if you are not
-creating a new test suite.
-
-To generate this file, from a command prompt at the root
-directory of the repo, run:
-
-```
-python tools/scripts/manifest.py
-```
-This `tools/scripts/manifest.py` needs python `html5lib` package.
-If you have not installed it yet, run:
-```
-pip install html5lib
-```
-
-On Mac OS X, python is installed with Xcode, but pip is not. Try
-```
-sudo easy_install pip
-```
-if pip is not already on your system.
-
-
-Running the tests requires that the test environment be activated as
-described above. The runner can be found at `/tools/runner/index.html`
-on the local server i.e.
+The runner can be found at `/tools/runner/index.html` on the local
+server i.e.
 
 ```
 http://web-platform.test:8000/tools/runner/index.html
 ```
 
-in the default configuration.
+in the default configuration. The first time you use this it has to
+generate a manifest of all tests. This may take some time, so please
+be patient.
 
 Publication
 ===========
@@ -142,17 +118,9 @@ node tools/scripts/id2path.js your-id
 Branches
 ========
 
-In the vast majority of cases the **only** branch that you should need
-to care about is `master`.
-
-There is another branch called `CR`. This is a strict subset of
-`master` that is limited to features that are found in the Candidate
-Recommendation version of the relevant specifications.
-
-If you see other branches in the repository, you can generally safely
-ignore them. Please note that branches prefixed with `temp/` are
-temporary branches and **can** get deleted at some point. So don't
-base any work off them unless you want to see your work destroyed.
+In the vast majority of cases the **only** upstream branch that you
+should need to care about is `master`. If you see other branches in
+the repository, you can generally safely ignore them.
 
 Contributing
 ============
@@ -172,16 +140,63 @@ The way to contribute is just as usual:
 * create a branch for your changes:
   `git checkout -b your-name/topic`;
 * make your changes;
-* push that to your repo;
+* Run the lint script described below
+* Commit locally push that to your repo;
 * and send in a pull request based on the above.
 
-Please make your pull requests either to `master` or to a feature
-branch (but not to `CR`).
+A lint is available to test for common mistakes in testcases. It can
+be run with:
+
+```
+./lint
+```
+
+in the root of the checkout. It is also run for every submitted pull
+request, and branches with lint errors will not get merged. In the
+unusual case that the lint is reporting an error for something that is
+essential to your test, there is a whitelist at
+`tools/lint/lint.whitelist` that may be updated.
+
+Adding command-line scripts ("tools" subdirs)
+---------------------------------------------
+
+Sometimes you may want to add a script to the repository that's meant
+to be used from the command line, not from a browser (e.g., a script
+for generating test files). If you want to ensure (e.g., or security
+reasons) that such scripts won't be handled by the HTTP server, but
+will instead only be usable from the command line, then place them in
+either:
+
+* the `tools` subdir at the root of the repository, or
+
+* the `tools` subdir at the root of any top-level directory in the
+  repository which contains the tests the script is meant to be used
+  with
+
+Any files in those `tools` directories won't be handled by the HTTP
+server; instead the server will return a 404 if a user navigates to
+the URL for a file within them.
+
+If you want to add a script for use with a particular set of tests but
+there isn't yet any `tools` subdir at the root of a top-level
+directory in the repository containing those tests, you can create a
+`tools` subdir at the root of that top-level directory and place your
+scripts there.
+
+For example, if you wanted to add a script for use with tests in the
+`notifications` directory, create the `notifications/tools` subdir and
+put your script there.
+
+Test Review
+===========
 
 We can sometimes take a little while to go through pull requests
 because we have to go through all the tests and ensure that they match
 the specification correctly. But we look at all of them, and take
 everything that we can.
+
+Getting Involved
+================
 
 If you wish to contribute actively, you're very welcome to join the
 public-test-infra@w3.org mailing list (low traffic) by
@@ -196,30 +211,8 @@ is [archived][ircarchive].
 [ircarchive]: http://krijnhoetmer.nl/irc-logs/testing/
 [mailarchive]: http://lists.w3.org/Archives/Public/public-test-infra/
 
-Adding command-line scripts ("tools" subdirs)
-----------------------------------------------------
-Sometimes you may want to add a script to the repository that's meant to be used from the
-command line, not from a browser (e.g., a script for generating test files). If you want to
-ensure (e.g., or security reasons) that such scripts won't be handled by the HTTP server,
-but will instead only be usable from the command line, then place them in either:
-
-* the `tools` subdir at the root of the repository, or
-* the `tools` subdir at the root of any top-level directory in the repository which
-  contains the tests the script is meant to be used with
-
-Any files in those `tools` directories won't be handled by the HTTP server; instead the
-server will return a 404 if a user navigates to the URL for a file within them.
-
-If you want to add a script for use with a particular set of tests but there isn't yet any
-`tools` subdir at the root of a top-level directory in the repository containing those
-tests, you can create a `tools` subdir at the root of that top-level directory and place
-your scripts there.
-
-For example, if you wanted to add a script for use with tests in the `notifications`
-directory, create the `notifications/tools` subdir and put your script there.
-
 Documentation
--------------
+=============
 
 * [How to write and review tests](http://testthewebforward.org/docs/)
 * [Documentation for the wptserve server](http://wptserve.readthedocs.org/en/latest/)
