+++ docs/submission-process.md
@@ -10,9 +10,17 @@ on, e.g. `git checkout -b topic-name`
 
 * Make your changes
 
-* Push that to your GitHub repository
+* Run the `lint` script in the root of your checkout to detect common
+  mistakes in test submissions. This will also be run after submission
+  and any errors will prevent your PR being accepted. If it detects an
+  error that forms an essential part of your test, edit the list of
+  exceptions stored in `tools/lint/lint.whitelist`.
 
-* Send in a pull request based on the above
+* Commit your changes.
+
+* Push your local branch to your GitHub repository
+
+* Using the GitHub UI create a PUll Request for your branch.
 
 * When you get review comments, make more commits to your branch to
   address the comments (**note**: Do *not* rewrite existing commits using
