+++ lint
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+import sys
+
+from tools.lint import lint
+
+sys.exit(0 if lint.main() == 0 else 1)
