+++ manifest
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+import sys
+
+from tools.manifest import update
+
+update.main()
