+++ serve
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+from tools.serve import serve
+
+serve.main()
