+++ tools/manifest/__init__.py
@@ -1,0 +1,5 @@
+import manifest
+import update
+import tree
+import item
+import sourcefile
