+++ tools/manifest/log.py
@@ -1,0 +1,8 @@
+import logging
+
+logging.basicConfig()
+logger = logging.getLogger("manifest")
+logger.setLevel(logging.DEBUG)
+
+def get_logger():
+    return logger
